refactor: 認証周りの関数を分離
